Migrated last function out of todoListMainContainer function in domManip to clean up file. Moved todoItemGenerator into its own file and imported'exported
